separate out paqa_latency_monotonic.c, which is failing
This allows the rest of paqa_latency.c to pass.
