+++ qa/paqa_latency.c
@@ -163,9 +163,10 @@ PaError paqaCheckLatency( PaStreamParameters *outputParamsPtr,
     streamInfo = Pa_GetStreamInfo( stream );
     printf("Stream info: inputLatency  = %g\n", streamInfo->inputLatency );
     printf("Stream info: outputLatency = %g\n", streamInfo->outputLatency );
+    QA_ASSERT_TRUE( "streamInfo->outputLatency should be > 0", (streamInfo->outputLatency > 0.0) );
 
     err = Pa_StartStream( stream );
-    if( err != paNoError ) goto error2;
+    if( err != paNoError ) goto error;
 
     printf("Play for %d seconds.\n", NUM_SECONDS );
     Pa_Sleep( NUM_SECONDS * 1000 );
@@ -176,180 +177,20 @@ PaError paqaCheckLatency( PaStreamParameters *outputParamsPtr,
     printf("  maxDeltaDacTime = %f\n", dataPtr->maxDeltaDacTime );
 
     err = Pa_StopStream( stream );
-    if( err != paNoError ) goto error2;
+    if( err != paNoError ) goto error;
 
     err = Pa_CloseStream( stream );
     Pa_Sleep( 1 * 1000 );
 
-
-    printf("-------------------------------------\n");
+    printf("-------------------------------------\n\n");
     return err;
-error2:
-    Pa_CloseStream( stream );
-error1:
-    printf("-------------------------------------\n");
-    return err;
-}
 
-
-/*******************************************************************/
-static int paqaNoopCallback( const void *inputBuffer, void *outputBuffer,
-                          unsigned long framesPerBuffer,
-                          const PaStreamCallbackTimeInfo* timeInfo,
-                          PaStreamCallbackFlags statusFlags,
-                          void *userData )
-{
-    (void)inputBuffer;
-    (void)outputBuffer;
-    (void)framesPerBuffer;
-    (void)timeInfo;
-    (void)statusFlags;
-    (void)userData;
-    return paContinue;
-}
-
-/*******************************************************************/
-static int paqaCheckMultipleSuggested( PaDeviceIndex deviceIndex, int isInput )
-{
-    int i;
-    int numLoops = 11;
-    PaError err;
-    PaStream *stream;
-    PaStreamParameters streamParameters;
-    const PaStreamInfo* streamInfo;
-    double lowLatency;
-    double highLatency;
-    double finalLatency;
-    double sampleRate = SAMPLE_RATE;
-    const PaDeviceInfo *pdi = Pa_GetDeviceInfo( deviceIndex );
-    double previousLatency = 0.0;
-    int numChannels = 1;
-    int atMaximumLatency = 0; /* Set to 1 when we reach the limit. */
-    double detectedMaximumLatency = 0.0;
-
-    printf("------------------------ paqaCheckMultipleSuggested - %s\n",
-           (isInput ? "INPUT" : "OUTPUT") );
-    if( isInput )
-    {
-        lowLatency = pdi->defaultLowInputLatency;
-        highLatency = pdi->defaultHighInputLatency;
-        numChannels = (pdi->maxInputChannels < 2) ? 1 : 2;
-    }
-    else
-    {
-        lowLatency = pdi->defaultLowOutputLatency;
-        highLatency = pdi->defaultHighOutputLatency;
-        numChannels = (pdi->maxOutputChannels < 2) ? 1 : 2;
-    }
-    streamParameters.channelCount = numChannels;
-    streamParameters.device = deviceIndex;
-    streamParameters.hostApiSpecificStreamInfo = NULL;
-    streamParameters.sampleFormat = paFloat32;
-    sampleRate = pdi->defaultSampleRate;
-
-    printf(" lowLatency   = %g\n", lowLatency );
-    printf(" highLatency  = %g\n", highLatency );
-    printf(" numChannels  = %d\n", numChannels );
-    printf(" sampleRate   = %g Hz\n", sampleRate );
-    printf(" samplePeriod = %e seconds\n", 1.0 / sampleRate );
-
-    if( highLatency < 0.001 )
-    {
-        numLoops = 2; /* Just test 0 and high. Don't divide by 0. */
-    }
-
-    for( i=0; i<numLoops; i++ )
-    {
-        streamParameters.suggestedLatency = highLatency * i /(numLoops - 1);
-        printf("   suggestedLatency[%2d] = %8.6f", i, streamParameters.suggestedLatency );
-
-        err = Pa_OpenStream(
-                            &stream,
-                            (isInput ? &streamParameters : NULL),
-                            (isInput ? NULL : &streamParameters),
-                            sampleRate,
-                            paFramesPerBufferUnspecified,
-                            paClipOff,      /* we won't output out of range samples so don't bother clipping them */
-                            paqaNoopCallback,
-                            NULL );
-        if( err != paNoError ) {
-            printf("\n");
-            goto error;
-        }
-
-        streamInfo = Pa_GetStreamInfo( stream );
-        // Get the latency from the streamInfo now because it will be invalid after the
-        // stream is closed.
-        if( isInput )
-        {
-            finalLatency = streamInfo->inputLatency;
-        }
-        else
-        {
-            finalLatency = streamInfo->outputLatency;
-        }
-        printf(", finalLatency = %8.6f", finalLatency );
-        printf(", (final - suggested) = %e\n", (finalLatency - streamParameters.suggestedLatency) );
-        err = Pa_CloseStream( stream );
-
-        QA_ASSERT_TRUE("Latency should be monotonically non-decreasing with suggested latency.",
-                        finalLatency >= previousLatency);
-        QA_ASSERT_TRUE("Latency should be > 0.0", finalLatency > 0.0);
-        if (atMaximumLatency == 0) {
-            /* If we get a lower value then we must be clipping at max latency. */
-            if (finalLatency < streamParameters.suggestedLatency) {
-                atMaximumLatency = 1;
-                detectedMaximumLatency = finalLatency;
-                printf("     detectedMaximumLatency = %8.6f\n", detectedMaximumLatency );
-            }
-        }
-        /* If we are not at maximum then we should be rounding up. */
-        if (atMaximumLatency == 0) {
-            QA_ASSERT_TRUE("Latency should be >= suggestedLatency",
-                           finalLatency >= streamParameters.suggestedLatency);
-        } else {
-            QA_ASSERT_TRUE("Latency should be == detectedMaximumLatency",
-                           finalLatency == detectedMaximumLatency);
-        }
-        previousLatency = finalLatency;
-    }
-
-    return 0;
 error:
-    return -1;
-}
-
-/*******************************************************************/
-static int paqaVerifySuggestedLatency( void )
-{
-    PaDeviceIndex id;
-    int result = 0;
-    const PaDeviceInfo *pdi;
-    int numDevices = Pa_GetDeviceCount();
-
-    printf("\n ------------------------ paqaVerifySuggestedLatency\n");
-    for( id=0; id<numDevices; id++ )            /* Iterate through all devices. */
-    {
-        pdi = Pa_GetDeviceInfo( id );
-        printf("\nUsing device #%d: '%s' (%s)\n", id, pdi->name, Pa_GetHostApiInfo(pdi->hostApi)->name);
-        if( pdi->maxOutputChannels > 0 )
-        {
-            if( paqaCheckMultipleSuggested( id, 0 /* isInput */ ) < 0 )
-            {
-                printf("OUTPUT CHECK FAILED !!! #%d: '%s'\n", id, pdi->name);
-                result -= 1;
-            }
-        }
-        if( pdi->maxInputChannels > 0 )
-        {
-            if( paqaCheckMultipleSuggested( id, 1 /* isInput */ ) < 0 )
-            {
-                printf("INPUT CHECK FAILED !!! #%d: '%s'\n", id, pdi->name);
-                result -= 1;
-            }
-        }
-    }
-    return result;
+    Pa_CloseStream( stream );
+    Pa_Sleep( 1 * 1000 );
+error1:
+    printf("-------------------------------------\n\n");
+    return err;
 }
 
 /*******************************************************************/
@@ -386,8 +227,6 @@ error:
     return -1;
 }
 
-
-
 /*******************************************************************/
 int main(void);
 int main(void)
@@ -415,8 +254,6 @@ int main(void)
     /* Run self tests. */
     if( (err = paqaVerifyDeviceInfoLatency()) < 0 ) goto error;
 
-    if( (err = paqaVerifySuggestedLatency()) < 0 ) goto error;
-
     outputParameters.device = Pa_GetDefaultOutputDevice(); /* default output device */
     if (outputParameters.device == paNoDevice) {
         fprintf(stderr,"Error: No default output device.\n");
@@ -436,7 +273,7 @@ int main(void)
     sampleRate = deviceInfo->defaultSampleRate;
     printf("Sample Rate for following tests: %g\n", sampleRate);
     outputParameters.hostApiSpecificStreamInfo = NULL;
-    printf("-------------------------------------\n");
+    printf("-------------------------------------\n\n");
 
     // Try to use a small buffer that is smaller than we think the device can handle.
     // Try to force combining multiple user buffers into a host buffer.
